requirements.txt
appdirs==1.4.4
autopep8==2.0.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
dnspython==2.2.1
frozendict==2.3.4
html5lib==1.1
idna==3.4
kafka-python==2.0.2
lxml==4.9.2
multitasking==0.0.11
newrelic==8.5.0
numpy==1.24.1
pandas==1.5.2
pycodestyle==2.10.0
pycparser==2.21
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.13
webencodings==0.5.1
yfinance==0.2.3